refactor: 전체 메인 화면에서 사용되는 state 객체를 별도의 패키지로 분리
