+++ feature/main/src/main/java/com/practice/main/main/MainScreenElements.kt
@@ -64,12 +64,12 @@ import com.practice.designsystem.modifier.drawBottomBorder
 import com.practice.designsystem.theme.BlindarTheme
 import com.practice.main.R
 import com.practice.main.main.state.MemoDialogElement
-import com.practice.main.main.state.UiMeal
-import com.practice.main.main.state.UiMeals
-import com.practice.main.main.state.UiMemo
-import com.practice.main.main.state.UiMenu
-import com.practice.main.main.state.UiNutrient
-import com.practice.main.main.state.UiSchedule
+import com.practice.main.state.UiMeal
+import com.practice.main.state.UiMeals
+import com.practice.main.state.UiMemo
+import com.practice.main.state.UiMenu
+import com.practice.main.state.UiNutrient
+import com.practice.main.state.UiSchedule
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
