+++ feature/main/src/main/java/com/practice/main/main/MainScreenViewModel.kt
@@ -23,15 +23,15 @@ import com.practice.firebase.BlindarUserStatus
 import com.practice.main.main.state.DailyData
 import com.practice.main.main.state.MainUiMode
 import com.practice.main.main.state.MainUiState
-import com.practice.main.main.state.UiMeals
-import com.practice.main.main.state.UiMemo
-import com.practice.main.main.state.UiMemos
-import com.practice.main.main.state.UiSchedules
-import com.practice.main.main.state.toMealUiState
-import com.practice.main.main.state.toMemo
-import com.practice.main.main.state.toUiMemo
 import com.practice.main.main.state.toUiMode
-import com.practice.main.main.state.toUiSchedule
+import com.practice.main.state.UiMeals
+import com.practice.main.state.UiMemo
+import com.practice.main.state.UiMemos
+import com.practice.main.state.UiSchedules
+import com.practice.main.state.toMealUiState
+import com.practice.main.state.toMemo
+import com.practice.main.state.toUiMemo
+import com.practice.main.state.toUiSchedule
 import com.practice.preferences.PreferencesRepository
 import com.practice.util.date.DateUtil
 import com.practice.work.BlindarWorkManager
