+++ feature/main/src/main/java/com/practice/main/main/daily/VerticalDailyMainScreen.kt
@@ -37,10 +37,10 @@ import com.practice.main.main.previewUiNutrients
 import com.practice.main.main.state.DailyData
 import com.practice.main.main.state.MainUiMode
 import com.practice.main.main.state.MainUiState
-import com.practice.main.main.state.UiMeal
-import com.practice.main.main.state.UiMeals
-import com.practice.main.main.state.UiMemos
-import com.practice.main.main.state.UiSchedules
+import com.practice.main.state.UiMeal
+import com.practice.main.state.UiMeals
+import com.practice.main.state.UiMemos
+import com.practice.main.state.UiSchedules
 import kotlin.math.absoluteValue
 
 @OptIn(ExperimentalFoundationApi::class)
