+++ feature/main/src/main/java/com/practice/main/state/StateUtil.kt
@@ -1,5 +1,6 @@
-package com.practice.main.main.state
+package com.practice.main.state
 
+import com.practice.main.main.state.MemoDialogElement
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
