+++ feature/main/src/main/java/com/practice/main/state/UiMeal.kt
@@ -1,4 +1,4 @@
-package com.practice.main.main.state
+package com.practice.main.state
 
 import com.hsk.ktx.date.Date
 import com.practice.domain.meal.Meal
