+++ feature/main/src/main/java/com/practice/main/state/UiMenu.kt
@@ -1,4 +1,4 @@
-package com.practice.main.main.state
+package com.practice.main.state
 
 data class UiMenu(
     val name: String,
